Fix CWE-682: prevent silent integer overflow in aggregate calculations
- ReactorService.totalOnlineOutputMw(): change from mapToInt().sum() (int) to
  mapToLong().sum() (long); return type updated from Integer to long
- IncidentService.totalDonuts(): change from mapToInt().sum() (int) to
  mapToLong().sum() (long); return type updated from int to long
- Both methods now accumulate via LongStream, which cannot silently wrap on
  realistic plant-scale inputs (max long ~9.2e18 MW / donuts)

+++ src/main/java/com/springfield/plant/service/IncidentService.java
@@ -48,9 +48,9 @@ public class IncidentService {
     }
 
     @Transactional(readOnly = true)
-    public int totalDonuts() {
+    public long totalDonuts() {
         return incidentRepository.findAll().stream()
-                .mapToInt(SafetyIncident::getDonutsConsumedDuringIncident)
+                .mapToLong(SafetyIncident::getDonutsConsumedDuringIncident)
                 .sum();
     }
